[AvUtil] Allocate memory by int bytes instead of ptr

+++ avutil/memory.go
@@ -15,7 +15,7 @@ import (
 )
 
 //Allocate a block of size bytes with alignment suitable for all memory accesses (including vectors if available on the CPU).
-func AvMalloc(s uintptr) unsafe.Pointer {
+func AvMalloc(s int) unsafe.Pointer {
 	return unsafe.Pointer(C.av_malloc(C.size_t(s)))
 }
